langgraph_checkpoint_dynamodb/tests/integration/test_localstack_verification.py
"""Verification tests to confirm integration tests run against LocalStack."""

import pytest
import boto3
from botocore.exceptions import ClientError

from langgraph_checkpoint_dynamodb import DynamoDBSaver

pytestmark = pytest.mark.integration


class TestLocalStackVerification:
    """Tests to verify LocalStack is actually running and being used."""

    def test_localstack_endpoint_verification(self, dynamodb_config):
        """Verify that we're connecting to LocalStack, not real AWS."""
        # Check endpoint URL
        endpoint_url = dynamodb_config.endpoint_url
        assert endpoint_url is not None
        assert (
            "localhost" in endpoint_url or "127.0.0.1" in endpoint_url
        ), f"Endpoint {endpoint_url} doesn't look like LocalStack (should be localhost/127.0.0.1)"
        # Should NOT be a real AWS endpoint
        assert (
            "amazonaws.com" not in endpoint_url
        ), f"Endpoint {endpoint_url} looks like real AWS, not LocalStack!"
        assert endpoint_url.startswith("http://") or endpoint_url.startswith(
            "https://"
        ), f"Endpoint {endpoint_url} should be a valid URL"

    def test_dynamodb_client_connects_to_localstack(self, dynamodb_config):
        """Verify DynamoDB client actually connects to LocalStack."""
        client = boto3.client(
            "dynamodb",
            endpoint_url=dynamodb_config.endpoint_url,
            region_name=dynamodb_config.region_name,
            aws_access_key_id=dynamodb_config.aws_access_key_id,
            aws_secret_access_key=dynamodb_config.aws_secret_access_key,
        )

        # Try to list tables - should work with LocalStack
        response = client.list_tables()
        assert "TableNames" in response

    def test_table_creation_in_localstack(self, checkpointer, dynamodb_config):
        """Verify table creation actually happens in LocalStack."""
        # The checkpointer fixture creates a table, verify it exists
        client = boto3.client(
            "dynamodb",
            endpoint_url=dynamodb_config.endpoint_url,
            region_name=dynamodb_config.region_name,
            aws_access_key_id=dynamodb_config.aws_access_key_id,
            aws_secret_access_key=dynamodb_config.aws_secret_access_key,
        )

        table_name = checkpointer.config.table_config.table_name

        # Verify table exists in LocalStack
        response = client.describe_table(TableName=table_name)
        assert response["Table"]["TableName"] == table_name
        assert response["Table"]["TableStatus"] == "ACTIVE"

        # Verify we're not hitting real AWS by checking region
        # LocalStack might not set this correctly, but we can check endpoint
        assert (
            "localhost" in dynamodb_config.endpoint_url
            or "127.0.0.1" in dynamodb_config.endpoint_url
        )

    def test_operation_fails_without_localstack_endpoint(self):
        """Verify that operations fail if endpoint is not set (safety check)."""
        from langgraph_checkpoint_dynamodb import DynamoDBConfig, DynamoDBTableConfig

        # Create config without endpoint (would try to hit real AWS)
        config = DynamoDBConfig(
            table_config=DynamoDBTableConfig(table_name="test-fail"),
            # No endpoint_url - would try real AWS
        )

        # This should fail because we don't have real AWS credentials in test environment
        # Note: We can't actually run this without credentials, but we can verify
        # the config doesn't have an endpoint
        assert config.endpoint_url is None, (
            "This test verifies that missing endpoint_url would fail. "
            "If endpoint_url is set, it means we're properly using LocalStack."
        )

    def test_checkpointer_uses_localstack_endpoint(self, checkpointer, dynamodb_config):
        """Verify checkpointer is configured with LocalStack endpoint."""
        # Check that checkpointer's config uses LocalStack endpoint
        checkpointer_config = checkpointer.config
        assert checkpointer_config.endpoint_url == dynamodb_config.endpoint_url
        assert (
            "localhost" in checkpointer_config.endpoint_url
            or "127.0.0.1" in checkpointer_config.endpoint_url
        )

        # Verify we can actually use it (proves it's connecting to LocalStack)
        from langchain_core.runnables import RunnableConfig

        test_config = RunnableConfig(
            configurable={"thread_id": "test_verification", "checkpoint_id": "test_1"}
        )
        checkpoint = {"id": "test_1"}
        metadata = {"step": 1}

        # This should work if LocalStack is running
        result = checkpointer.put(test_config, checkpoint, metadata, {})
        assert result is not None

        # Verify we can read it back (confirms LocalStack persistence)
        result_tuple = checkpointer.get_tuple(test_config)
        assert result_tuple is not None
        assert result_tuple.checkpoint["id"] == "test_1"
